+++ includes/minirt.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minirt.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/08 20:57:46 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/16 20:18:19 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/22 21:37:18 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,12 +20,12 @@
 # include <fcntl.h>
 # include <mlx.h>
 # include <math.h>
-// # include "libft/libft.h"
 # include "libft.h"
 # include "constants.h"
 # include "dtos.h"
 # include "functions/coloring.h"
 # include "functions/parser.h"
+# include "functions/render.h"
 # include "functions/utils.h"
 
 #endif