+++ src/coloring/set_color_matrix.c
@@ -3,14 +3,15 @@
 /*                                                        :::      ::::::::   */
 /*   set_color_matrix.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/16 19:04:24 by jvacaris          #+#    #+#             */
-/*   Updated: 2022/03/21 17:41:56 by jvacaris         ###   ########.fr       */
+/*   Updated: 2022/03/22 21:16:49 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
+
 
 /*
 ?		In this file, we'll fill each matrix pixel with its color.
@@ -30,27 +31,17 @@ TODO	the pixels.
 *		*	Get a normal vector of the collision point so ve can calculate ray angles.
 *		*	Apply the color mixing (color of the figure + color of the lights).
 */
-t_colors	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
+int	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 {
-	t_colors	**c_matrix;
+	int			**c_matrix;
 	int			x;
 	int			y;
 
-//! ~~ Testing started ~~~~~~~~~~
-//	/*	
-	void	*gnrl_ptr;
-	void	*win_ptr;
-
-	gnrl_ptr = mlx_init();
-	win_ptr = mlx_new_window(gnrl_ptr, HORIZ_RESOLUTION, RESOLUTION, "miniRT test");
-//	*/
-//! ~~~ Testing ended ~~~~~~~~~~~
-
-	c_matrix = malloc(sizeof(t_coords *) * HORIZ_RESOLUTION);
+	c_matrix = malloc(sizeof(int *) * HORIZ_RESOLUTION);
 	x = 0;
 	while (x < HORIZ_RESOLUTION)
 	{
-		c_matrix[x] = malloc(sizeof(t_coords) * RESOLUTION);
+		c_matrix[x] = malloc(sizeof(int) * RESOLUTION);
 		x++;
 	}
 	y = 0;
@@ -61,25 +52,15 @@ t_colors	**set_color_matrix(t_itemlist *items, t_coords **v_matrix)
 		{
 			if (check4collisions(v_matrix[x][y], items))
 			{
-
-			//! ~~ Testing started ~~~~~~~~~~
-//	/*
-				mlx_pixel_put (gnrl_ptr, win_ptr, x, y, 16777215);
-//	*/
-			//! ~~~ Testing ended ~~~~~~~~~~~
-
+				c_matrix[x][y] = 0x00FFFFFF;
 				//TODO	Here we'll calculate collision point & get its color.
 			}
+			else
+				c_matrix[x][y] = 0x0;
 			x++;
 		}
 		y++;
 	}
-
-//! ~~ Testing started ~~~~~~~~~~
-//	/*
-	mlx_loop(gnrl_ptr);
-//	*/
-//! ~~~ Testing ended ~~~~~~~~~~~
-
 	return (c_matrix);
 }
+